Change migration file for schools to set amount owed as float

+++ db/migrate/20150112212957_create_schools.rb
@@ -4,7 +4,7 @@ class CreateSchools < ActiveRecord::Migration
       t.string :dbn
       t.string :school
       t.integer :total_enrollment
-      t.integer :amount_owed
+      t.float :amount_owed
 
       t.timestamps
     end
